requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
tortoise-orm==0.21.7
aiosqlite==0.20.0
coinbase-advanced-py==1.8.2
APScheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.2

# dev / tests
pytest==8.3.3
pytest-asyncio==0.24.0
